backend/requirements.txt
fastapi
uvicorn
pillow
torch
accelerate
bitsandbytes
pymupdf
opencv-python
numpy
python-multipart
safetensors
pydantic
Pillow
transformers>=4.39.0
protobuf